test(e2e): move the files around
