+++ e2e/start/basic/app/routes/server-fns.tsx
@@ -5,7 +5,7 @@ import { ConsistentServerFnCalls } from './-server-fns/consistent-fn-calls'
 import { MultipartServerFnCall } from './-server-fns/multipart-formdata-fn-call'
 import { AllowServerFnReturnNull } from './-server-fns/allow-fn-return-null'
 import { SerializeFormDataFnCall } from './-server-fns/serialize-formdata-fn-call'
-import { SubmitFormDataFn } from './-server-fns/submit-formdata-fn'
+import { SubmitPostFormDataFn } from './-server-fns/submit-post-formdata-fn'
 
 export const Route = createFileRoute('/server-fns')({
   component: RouteComponent,
@@ -18,7 +18,7 @@ function RouteComponent() {
       <MultipartServerFnCall />
       <AllowServerFnReturnNull />
       <SerializeFormDataFnCall />
-      <SubmitFormDataFn />
+      <SubmitPostFormDataFn />
     </>
   )
 }
